add: QMK to data.json
I kept the description short and sweet in order to play nicely in the description field.

The assets from https://hacktoberfest.qmk.fm/ can be found here:

https://github.com/qmk/hacktoberfest/tree/master/img

It does not seem have anything that looks great on the image element however I believe the `social.jpg` assets has a nice contrast to the white background of the site.